fix : view 디렉토리 변경
