+++ back/app/views_item/views_item.py
@@ -1,58 +1,14 @@
-from django.shortcuts import get_object_or_404, get_list_or_404
-from django.db.models import Q
-from django.db.models import F
-
 from rest_framework import status
-from rest_framework import viewsets
 from rest_framework.response import Response
-from rest_framework.decorators import api_view, action
-from rest_framework.authtoken.models import Token
-from rest_framework.serializers import Serializer
+from rest_framework.decorators import api_view
 
-from .serializers import *
-from .models import *
-from .middleware import validation_token
+from ..serializers import *
+from ..models import *
+from ..middleware import validation_token
 
 import time
 import random
 
-@api_view(['POST'])
-def registration_view(request):
-    
-    if request.method == 'POST':
-        serializer = RegisterationSerializer(data=request.data)
-        data = {}
-        if serializer.is_valid():
-            user = serializer.save()
-            data['response'] = "Registerd succesfully"
-            data['username'] = user.username
-            token = Token.objects.get(user=user).key
-            data['token'] = token 
-
-        else:
-            data = serializer.errors
-        return Response(data)
-
-
-# /users
-@api_view(['GET'])
-def users(request):
-    if request.method == 'GET':
-        users = User.objects.all()
-        serializer = UserSerializer(users, many=True)
-        return Response(serializer.data)
-
-
-# /users/<username>
-@api_view(['GET'])
-def user_detail(request, username):
-    if request.method == 'GET':
-        user = User.objects.get(username=username)
-        serializer = UserSerializer(user)
-        return Response(serializer.data)
-
-
-# /items
 @api_view(['GET'])
 def items(request):
     
@@ -78,7 +34,6 @@ def items(request):
         return Response(serializer.data, status=status.HTTP_200_OK) 
 
 
-# items/<int:pk>        
 @api_view(['GET'])
 def item_detail(request, pk):
     if request.method == 'GET':
@@ -108,7 +63,6 @@ def item_detail(request, pk):
         return Response(item_detail, status=status.HTTP_200_OK)
 
 
-# /wishlist/<user_id>
 @api_view(['GET'])
 def wanted_item(request, id):
     if request.method == 'GET':
@@ -124,62 +78,6 @@ def wanted_item(request, id):
 
         serializer = ItemSerializer(item_list, many=True)
         return Response(serializer.data)
-
-
-@api_view(['GET'])
-def item_search(request):
-    start = time.time()
-
-    if request.method == 'GET':
-
-        if "invalid_token" in validation_token(request).data:
-            return Response(validation_token(request).data, status=status.HTTP_401_UNAUTHORIZED)
-
-        try:
-            keyword = request.query_params.get('keyword', None)
-
-            '''
-            __icontain -> 대소문자 알아서 비교
-            JorDan으로 검색을 해도, Jordan, JORDAN, jordan 등을 모두 검색할 수 있다.
-            '''
-            items = Item.objects.filter(
-                Q(title__icontains=keyword) |
-                Q(content__icontains=keyword)
-            )          
-
-            print(str(items.query))
-
-            # 검색어에 해당하는 데이터가 없는 경우  
-            if not items[0]:
-                return Response({"Error message":"No Item for your request"}, status=status.HTTP_404_NOT_FOUND)
-
-        except:
-            return Response({"Error message":"No Item for your request"}, status=status.HTTP_404_NOT_FOUND)
-        
-        try:
-            offset = int(request.query_params.get('offset'))
-            limit = int(request.query_params.get('limit'))
-            items = items[offset:limit+offset]
-            serializer = ItemSerializer(items, many=True)
-
-        except:
-            return Response({"Error message":"Wrong offset or limit"}, status=status.HTTP_404_NOT_FOUND)
-
-        # limit값은 최대 20~30일 것 -> 성능에 크게 문제 없음
-        for data in serializer.data:
-            username = User.objects.get(id=data['sell_username']).username
-            data['sell_user_id'] = data['sell_username']
-            data['sell_username'] = username
-
-        print("time :", time.time() - start)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-
-@api_view(['GET', 'POST'])
-def hello_world(request):
-    if request.method == 'POST':
-        return Response({"message": "Got some data!", "data": request.data})
-    return Response({"message": "Hello, world!"})
 
 
 # 배포용은 아니다. 나중에 수정을 해야한다.
